backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.9.1
camelcase==0.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
crispy-bootstrap4==2025.6
Django==5.2.4
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-extensions==4.1
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
fastapi==0.124.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
mysqlclient==2.2.7
ollama==0.6.1
packaging==25.0
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.0
uvicorn==0.38.0
